requirements.txt
# AI/ML APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Text Processing
markitdown>=0.1.0
beautifulsoup4>=4.12.0
ebooklib>=0.18

# Web UI
streamlit>=1.50.0

# Image Processing
Pillow>=10.0.0

# Video Processing
moviepy>=1.0.3

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
